requirements.txt
aiohttp==3.8.0
aiopg==1.3.3
aiosignal==1.2.0
alembic==1.7.4
async-timeout==4.0.0
asyncpg==0.24.0
attrs==21.2.0
charset-normalizer==2.0.7
frozenlist==1.2.0
gino==1.0.1
idna==3.3
Mako==1.1.5
MarkupSafe==2.0.1
multidict==5.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dotenv==0.19.1
PyYAML==6.0
SQLAlchemy==1.3.24
typing-extensions==3.10.0.2
Werkzeug==2.0.2
yarl==1.7.2
